+++ includes/class-stream-merger.php
@@ -26,6 +26,7 @@ use Newspack_Nodes\Config;
 use Newspack_Nodes\Core;
 use Newspack_Nodes\Message;
 use Newspack_Nodes\Node;
+use Newspack_Nodes\Node_Names;
 use Newspack_Nodes\Partition;
 
 if ( ! \defined( 'ABSPATH' ) ) {
@@ -252,7 +253,7 @@ class StreamMerger extends Node {
 	}
 
 	public function start_periodic_tick(): void {
-		$router = Core::node( '_router' );
+		$router = Core::node( Node_Names::ROUTER );
 		if ( null === $router ) {
 			Core::print_less_often( 'StreamMerger::start_periodic_tick: no _router; periodic tick disabled' );
 			return;